Consolidate aperture extraction functions
